requirements.txt
fastapi==0.111.1
uvicorn==0.30.3
requests==2.32.3
pytest==8.3.1
Faker~=26.0.0
pydantic~=2.8.2
jsonschema~=4.23.0
fastapi-pagination
python-dotenv~=1.0.1
psycopg2-binary==2.9.9
SQLAlchemy==2.0.31
sqlmodel